requirements.txt
pyyaml>=6.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
click>=8.1.0
tenacity>=8.2.0
